price_calculator/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
jinja2>=3.1.0
python-multipart>=0.0.6